Alternate ANES preprocessing script
I've been rewriting the "Exploring the ANES" script from scratch to
better grok how it works, the associated ANES variables, etc.

There are various lines, marked with NOTE, indicating primary
differences. The biggest is how 99s are handled in PID/ideology
(should they be recoded to 4 or to NA), which affects the sorting
variable downstream.